refactor(core): split tjpgdec into own crate
[no changelog]
